refactor: split RdmParameterData impls into separate files, added tests
